+++ components/zcash_address/src/lib.rs
@@ -5,8 +5,7 @@
 //!
 //! - [`ZcashAddress`] can be parsed from, and encoded to, strings.
 //! - [`ZcashAddress::convert`] or [`ZcashAddress::convert_if_network`] can be used to
-//!   convert a parsed address into custom types that implement the [`TryFromAddress`] or
-//!   [`TryFromRawAddress`] traits.
+//!   convert a parsed address into custom types that implement the [`TryFromAddress`] trait.
 //! - Custom types can be converted into a [`ZcashAddress`] via its implementation of the
 //!   [`ToAddress`] trait.
 //!
@@ -63,7 +62,7 @@
 //! use std::ffi::{CStr, c_char, c_void};
 //! use std::ptr;
 //!
-//! use zcash_address::{ConversionError, TryFromRawAddress, ZcashAddress};
+//! use zcash_address::{ConversionError, ZcashAddress};
 //! use zcash_protocol::consensus::NetworkType;
 //!
 //! // Functions that return a pointer to a heap-allocated address of the given kind in
@@ -75,10 +74,10 @@
 //!
 //! struct ParsedAddress(*mut c_void);
 //!
-//! impl TryFromRawAddress for ParsedAddress {
+//! impl TryFromAddress for ParsedAddress {
 //!     type Error = &'static str;
 //!
-//!     fn try_from_raw_sapling(
+//!     fn try_from_sapling(
 //!         data: [u8; 43],
 //!     ) -> Result<Self, ConversionError<Self::Error>> {
 //!         let parsed = unsafe { addr_from_sapling(data[..].as_ptr()) };
@@ -89,7 +88,7 @@
 //!         }
 //!     }
 //!
-//!     fn try_from_raw_transparent_p2pkh(
+//!     fn try_from_transparent_p2pkh(
 //!         data: [u8; 20],
 //!     ) -> Result<Self, ConversionError<Self::Error>> {
 //!         let parsed = unsafe { addr_from_transparent_p2pkh(data[..].as_ptr()) };
@@ -116,7 +115,7 @@
 //!     match addr.convert_if_network::<ParsedAddress>(NetworkType::Main) {
 //!         Ok(parsed) => parsed.0,
 //!         Err(e) => {
-//!             // We didn't implement all of the methods of `TryFromRawAddress`, so if an
+//!             // We didn't implement all of the methods of `TryFromAddress`, so if an
 //!             // address with one of those kinds is parsed, it will result in an error
 //!             // here that should be passed back across the FFI.
 //!             ptr::null_mut()
@@ -146,9 +145,8 @@ mod kind;
 #[cfg(any(test, feature = "test-dependencies"))]
 pub mod test_vectors;
 
-pub use convert::{
-    ConversionError, Converter, ToAddress, TryFromAddress, TryFromRawAddress, UnsupportedAddress,
-};
+use convert::Converter;
+pub use convert::{ConversionError, ToAddress, TryFromAddress, UnsupportedAddress};
 pub use encoding::ParseError;
 pub use kind::unified;
 use kind::unified::Receiver;
@@ -256,7 +254,7 @@ impl ZcashAddress {
     /// [`encode`]: Self::encode
     /// [`Display` implementation]: core::fmt::Display
     /// [`address.to_string()`]: alloc::string::ToString
-    pub fn convert_if_network<T: TryFromRawAddress>(
+    pub fn convert_if_network<T: TryFromAddress>(
         self,
         net: NetworkType,
     ) -> Result<T, ConversionError<T::Error>> {
@@ -267,14 +265,14 @@ impl ZcashAddress {
             network_matches || (self.net == NetworkType::Test && net == NetworkType::Regtest);
 
         match self.kind {
-            AddressKind::Sprout(data) if regtest_exception => T::try_from_raw_sprout(data),
-            AddressKind::Sapling(data) if network_matches => T::try_from_raw_sapling(data),
-            AddressKind::Unified(data) if network_matches => T::try_from_raw_unified(data),
+            AddressKind::Sprout(data) if regtest_exception => T::try_from_sprout(net, data),
+            AddressKind::Sapling(data) if network_matches => T::try_from_sapling(net, data),
+            AddressKind::Unified(data) if network_matches => T::try_from_unified(net, data),
             AddressKind::P2pkh(data) if regtest_exception => {
-                T::try_from_raw_transparent_p2pkh(data)
+                T::try_from_transparent_p2pkh(net, data)
             }
-            AddressKind::P2sh(data) if regtest_exception => T::try_from_raw_transparent_p2sh(data),
-            AddressKind::Tex(data) if network_matches => T::try_from_raw_tex(data),
+            AddressKind::P2sh(data) if regtest_exception => T::try_from_transparent_p2sh(net, data),
+            AddressKind::Tex(data) if network_matches => T::try_from_tex(net, data),
             _ => Err(ConversionError::IncorrectNetwork {
                 expected: net,
                 actual: self.net,