+++ zcash_transparent/CHANGELOG.md
@@ -8,7 +8,7 @@ and this library adheres to Rust's notion of
 ## [Unreleased]
 
 ### Changed
-- Updated to `zcash_address 0.8.0`
+- Migrated to `zcash_address 0.8`.
 
 ## [0.2.3] - 2025-04-04